fix: rename incorrectly named file
